+++ sciencebeam_trainer_grobid_tools/annotation/segmentation_annotator.py
