+++ sciencebeam_trainer_grobid_tools/auto_annotate_reference.py
@@ -30,11 +30,11 @@ from .auto_annotate_utils import (
     AnnotatorConfig
 )
 
-from .structured_document.simple_matching_annotator import (
+from .annotation.simple_matching_annotator import (
     SimpleMatchingAnnotator
 )
 
-from .structured_document.reference_annotator import (
+from .annotation.reference_annotator import (
     DEFAULT_IDNO_PREFIX_REGEX,
     ReferenceAnnotatorConfig,
     ReferencePostProcessingAnnotator
